// src/features/pix/redux/models/response/findPixQrCodeInfoResponse.ts
export interface FindPixQrCodeInfoResponse {
  searchProtocol?: number;
  hash?: string;
  receiverBank?: string;
  receiverBankBranch?: string;
  receiverBankAccount?: string;
  receiverBankAccountDigit?: string;
  receiverAccountType?: string;
  type?: string;
  status?: string;
  pixKeyValue?: string;
  externalIdentifier?: string;
  receiverName?: string;
  receiverTaxNumber?: string;
  description?: string;
  payerName?: string;
  payerTaxNumber?: string;
  zipCode?: string;
  city?: string;
  originalValue?: string;
  expirationDate?: string;
  originalCreationDate?: string;
  url?: string;
  reusable?: string;
  UF?: string;
  rebateValue?: string;
  discountValue?: string;
  interestValue?: string;
  finalValue?: string;
  fineValue?: string;
  dueDate?: string;
  daysAfterDueDate?: string;
  paymentDate?: Date;
  categoryCode?: string;
  additionalData?: string;
  payerRequest?: string;
  expirationQrCode?: string;
  receiverTradingName?: string;
  receiverPublicPlace?: string;
  ispbPss?: string;
  changeType?: string;
  withdrawTransactionValue?: string;
  transactionChangeType?: string;
  withdrawIspbPss?: string;
  withdrawAgentModality?: string;
  purchaseWithChangeValue?: string;
  purchaseWithChangeTransactionChangeType?: string;
  purchaseWithChangeIspbPss?: string;
  purchaseWithChangeAgentModality?: string;
}